Refactor: move monkey_patch to patches/awesome_print
