+++ lib/run_loop.rb
@@ -15,7 +15,7 @@ require 'run_loop/device'
 require 'run_loop/instruments'
 require 'run_loop/lipo'
 require 'run_loop/host_cache'
-require 'run_loop/monkey_patch'
+require 'run_loop/patches/awesome_print'
 require 'run_loop/simctl/bridge'
 
 module RunLoop
