+++ spec/integration/patches/awesome_print_spec.rb
